+++ History.md
@@ -1,9 +1,14 @@
 # Version 3.2.0
 Release date: unreleased
 
+### Changed
+
+* Ruby 2.3.0+ is now required
+
 ### Added
 
 * New global configuration `default_set_options` used in `Capybara::Node::Element#set` as default `options` hash [Champier Cyril]
+* `execute_javascript' and `evaluate_javascript` can now be called on elements to run the JS in the context of the element [Thomas Walpole]
 
 # Version 3.1.0
 Release date: 2018-05-10